Turnitin should control visibility of gradebook items
Plugins in Moodle can declare they wish to control the visibility of items in the gradebook.

This patch puts the Turnitin tool plugin in control of the visibility of it's own gradebook items.

Users should no longer be able to edit the visibility of it's grade items via the gradebook.

Grades will be set to become avaliable from the highest post date, unless the Turnitin activity is hidden in which case the grade items will be hidden.