Fix division by zero error
When a part of an assignment has max grade of zero and we display the grade as a percentage, we get a division by zero error that results in an error 500. We need to check that maxmarks is greater than zero before doing the division. Otherwise, just show the submission grade instead of a percentage.